+++ yodl/src/args/argsnargs.c
@@ -1,6 +1,6 @@
 #include "args.ih"
 
-unsigned args_nArgs()
+size_t args_nArgs()
 {
     return args.d_argc - args.d_optind;
 }
