+++ yodl/src/lexer/lmaxreplacementsexceeded.c
@@ -1,6 +1,6 @@
 #include "lexer.ih"
 
-void l_max_replacements_exceeded(unsigned max)
+void l_max_replacements_exceeded(size_t max)
 {
     message_show(MSG_CRIT);
     message("Max number of macro/subst replacements (%u) exceeded\n", max);
