+++ yodl/src/lexer/lmedialineno.c
@@ -1,6 +1,6 @@
 #include "lexer.ih"
 
-unsigned l_media_lineno(register Media *mp)
+size_t l_media_lineno(register Media *mp)
 {
     return mp->d_lineno;
 }
