+++ yodl/src/lines/linesat.c
@@ -1,6 +1,6 @@
 #include "lines.h"
 
-char const *lines_at(register Lines *lines, unsigned idx)
+char const *lines_at(register Lines *lines, size_t idx)
 {
     return idx >= lines->d_size ? 0 : lines->d_str[idx];
 }
