+++ yodl/src/lines/linesformat.c
@@ -3,7 +3,7 @@
 void lines_format(Lines *lines, char const *fmt, ...)
 {
     va_list list;
-    unsigned n;
+    size_t n;
 
     va_start(list, fmt);
     lines_grab(lines, string_vformat(&n, fmt, list));
