+++ yodl/src/parser/parserwslevel.c
@@ -1,6 +1,6 @@
 #include "parser.ih"
 
-unsigned parser_ws_level(register Parser *pp)
+size_t parser_ws_level(register Parser *pp)
 {
     return pp->d_ws_level;
 }
